update to the latest version of HACD (hierarchical approximate convex decomposition, http://sourceforge.net/projects/hacd) Thanks to Khaled Mammou for the fix.

+++ Extras/HACD/hacdVector.h
@@ -21,7 +21,7 @@
 
 namespace HACD
 {
-    typedef float Real;
+    typedef double Real;
 	//!	Vector dim 3.
 	template < typename T > class Vec3
 	{
